refactor(project): components in tour page
