+++ src/components/Carousel/Carousel.jsx
@@ -1,0 +1,12 @@
+/** @format */
+
+import React, { Component } from 'react'
+import './Carousel.css'
+
+class Carousel extends Component {
+    render() {
+        return ;
+    }
+}
+
+export default Carousel
