+++ src/components/Carousel/CarouselHeadline/index.jsx
@@ -1,0 +1,4 @@
+/** @format */
+
+import CarouselHeadline from './CarouselHeadline'
+export default CarouselHeadline
