+++ src/components/Carousel/CarouselLocations/CarouselLocations.jsx
