+++ src/components/Carousel/CarouselLocations/index.jsx
@@ -1,0 +1,4 @@
+/** @format */
+
+import CarouselLocations from './CarouselLocations'
+export default CarouselLocations
