+++ src/components/Carousel/CarouselTour/index.jsx
@@ -1,0 +1,4 @@
+/** @format */
+
+import CarouselTour from './CarouselTour'
+export default CarouselTour
