+++ src/components/Carousel/index.jsx
@@ -1,0 +1,4 @@
+/** @format */
+
+import Carousel from './Carousel'
+export default Carousel
