+++ src/components/Heading/Heading.jsx
@@ -1,0 +1,12 @@
+/** @format */
+
+import React, { Component } from 'react'
+import './Heading.css'
+
+class Heading extends Component {
+    render() {
+        return ;
+    }
+}
+
+export default Heading
