+++ src/components/Heading/index.jsx
@@ -1,0 +1,4 @@
+/** @format */
+
+import Heading from './Heading'
+export default Heading
