--- src/components/Sections/AboutTravel/Components/LocationsCarousel/index.jsx
@@ -1,4 +1,0 @@
-/** @format */
-
-import LocationsCarousel from './LocationsCarousel'
-export default LocationsCarousel
