+++ src/components/Sections/Article/index.jsx
@@ -1,0 +1,4 @@
+/** @format */
+
+import Article from './Article'
+export default Article
