+++ src/components/Sections/Headline/HeadlineHome/index.jsx
@@ -1,0 +1,4 @@
+/** @format */
+
+import HeadlineHome from './HeadlineHome'
+export default HeadlineHome
