+++ src/components/Sections/Headline/HeadlineTour/HeadlineTour.css
@@ -1,0 +1,64 @@
+.contact__card {
+    color: #000;
+    padding: 30px;
+    margin-left: auto;
+    margin-right: auto;
+    cursor: pointer;
+    border: 2px solid transparent;
+    background: rgba(255, 255, 255, 0.3);
+    column-gap: 2rem;
+}
+
+.contact__subtitle {
+    width: 100%;
+}
+
+.subComponent-lg {
+    padding: 80px 0px;
+}
+
+
+.contact__card:hover {
+    border-style: solid;
+    border-width: 2px;
+    border-color: #e0e0e0;
+}
+
+.contact__title {
+    font-family: 'Poppins', sans-serif;
+    font-weight: 900;
+}
+
+.contact__form {
+    height: 100%;
+}
+
+.contact__entry {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    justify-content: center;
+    align-items: center;
+    row-gap: 1rem;
+}
+
+.contact__input {
+    width: 60%;
+    border-radius: 5px;
+    padding: 0.5rem;
+    border: 1px solid gray;
+}
+
+.contact__button {
+    width: 60%;
+}
+
+
+#contactBody {
+    /* background: #98d3a6; */
+    color: rgba(0, 0, 0, 0.7);
+}
+
+.headerTitle p {
+    color: rgba(0, 0, 0, 0.5);
+}
