+++ src/components/Sections/Headline/HeadlineTour/index.jsx
@@ -1,0 +1,4 @@
+/** @format */
+
+import HeadlineTour from './HeadlineTour'
+export default HeadlineTour
