+++ src/components/Sections/Reviews/Reviews.jsx
@@ -1,0 +1,12 @@
+/** @format */
+
+import React, { Component } from 'react'
+import './Reviews.css'
+
+class Reviews extends Component {
+    render() {
+        return ;
+    }
+}
+
+export default Reviews
