+++ src/components/Sections/Reviews/index.jsx
@@ -1,0 +1,4 @@
+/** @format */
+
+import Reviews from './Reviews'
+export default Reviews
